% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust-stat-chanperm.R
\name{channel_perm}
\alias{channel_perm}
\title{Clustering statistics with channel-wise permutations}
\usage{
channel_perm(
  dist_list,
  B,
  maxclust = NULL,
  nclust = NULL,
  strat.var = NULL,
  stat = c("ASW", "CH"),
  c_weight = rep(1, length(dist_list))
)
}
\arguments{
\item{dist_list}{list of dissimilarities in either format \code{'dist'} or as a symmetric \code{n x n} matrix.}

\item{B}{number of permutations}

\item{maxclust}{maximum number of clusters}

\item{nclust}{number of clusters. Ignored if maxclust is specified.}

\item{strat.var}{optional factor of length n by which the permutation is stratified.}

\item{stat}{vector of strings with cluster statistics to be calculated. Possible stats: 
\code{"ASW"} Average silhouette width.
\code{"CH"} Calinsky Harabasz index on squared distances.}

\item{c_weight}{Optional vector of numerics with channel weights.}
}
\value{
Object of class "ch_perm.stats".
}
\description{
Calculates clustering statistcs for the clustering of the (weighted) sum of a list of dissimilarity matrices. 
Performs (stratified) permutations and calculates clustering statistics for permuted  datasets.
}
\details{
...
}
\examples{
p <- channel_perm(d_imp_UD2,B = 50, maxclust = 15,strat.var = as.vector(TraMineR::seqlength(seq_mis[[1]])),c_weight = weigth, stat="ASW")
}
