% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{disthclustCBI}
\alias{disthclustCBI}
\title{Clusterwise cluster stability - Weighted 
Modification \code{\link{fpc::clusterboot}} to weight channels in each permutation}
\usage{
disthclustCBI(dmatrix, k, cut = "number", method, weights = NULL,
  noisecut = 0, ...)
}
\arguments{
\item{data}{n x p dataframe or dissimilarity matrix (if distances = TRUE) or list of dissimilarity matrices (if multichannel = TRUE)}

\item{indvec}{optional vector of indices if data is a dissimilarity matrix or  \code{'dist'} object repressenting only unique observations.}

\item{dt.m}{optional list of dataframe with data needed to weight channels (if multichannel = TRUE)}
}
\description{
Clusterwise cluster stability - Weighted 
Modification \code{\link{fpc::clusterboot}} to weight channels in each permutation
}

