% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred-strength-dist.R
\name{pred.strength.dist}
\alias{pred.strength.dist}
\title{Modification of \code{\link{prediction.strength\{fpc\}}} that allows input format to be a dissimilarity matrix and clustering method to be hierachical.}
\usage{
pred.strength.dist(
  dist,
  Gmin = 2,
  Gmax = 10,
  M = 50,
  cutoff = 0.8,
  clustermethod = c("ward.D2", "average"),
  class.method = c("ward.D2", "average", "pam"),
  indx = NULL,
  weights = NULL,
  mc.cores = NULL
)
}
\arguments{
\item{dist}{distance matrix or object of class \code{'dist'}.}
}
\value{
object of class predstr
}
\description{
Modification of \code{\link{prediction.strength\{fpc\}}} that allows input format to be a dissimilarity matrix and clustering method to be hierachical.
}
