% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdistmc.R
\name{seqdistmc2}
\alias{seqdistmc2}
\title{Dissimilarity between multichannel sequence object - save combined sub.matrix
Modification \code{\link{TraMineR::seqdistmc}} to save combined sub.matrix}
\usage{
seqdistmc2(channels, method, norm = FALSE, indel = 1, sm = NULL,
  with.missing = FALSE, full.matrix = TRUE, link = "sum", cval = 2,
  miss.cost = 2, cweight = NULL)
}
\arguments{
\item{channels}{list of stslist objects}

\item{sm}{}

\item{cweight}{}
}
\description{
Dissimilarity between multichannel sequence object - save combined sub.matrix
Modification \code{\link{TraMineR::seqdistmc}} to save combined sub.matrix
}

