% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri-ineq.R
\name{tri.ineq2}
\alias{tri.ineq2}
\title{Modification of \code{\link{tri.ineq\{fossil\}}} that gives magnitude and indices.}
\usage{
tri.ineq2(d)
}
\arguments{
\item{d}{distance matrix}
}
\value{
list containing \code{ineq} and \code{arr.indices}.
}
\description{
Modification of \code{\link{tri.ineq\{fossil\}}} that gives magnitude and indices.
}
