% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clusterstats.R
\name{wcCmpCluster2}
\alias{wcCmpCluster2}
\title{Modification of \code{\link{wcCmpCluster\{WeightedCluster\}}} that distinguishes hclust(... ,method="ward.D") and hclust(... ,method="ward.D2").}
\usage{
wcCmpCluster2(diss, weights = NULL, maxcluster, method = "all",
  pam.combine = TRUE)
}
\description{
See \code{\link{wcCmpCluster\{WeightedCluster\}}}
}

